[DEBUG] integer type was not loose enough

+++ initializeObservation.m
@@ -36,7 +36,7 @@ binR = (r(1:end-1)+r(2:end))/2;
 binR = [-dr/2 binR binR(end)+dr/2];
 
 % initialize matrix of observations
-energy = zeros(Nr,Npsi,Nt,1+~acoustics,'uint16');
+energy = zeros(Nr,Npsi,Nt,1+~acoustics,'uint32');
 Ec = zeros(Nt,2);
 
 % energy in a small volume of the domain
